Fix issue with export and $ not working if variable name begin with _
