+++ src/export_utils.c
@@ -33,11 +33,11 @@ bool	checkvalidenv(char *arg)
 	int	i;
 
 	i = 0;
-	if (ft_isalpha(arg[0]) == 0)
+	if (ft_isalpha(arg[0]) == 0 && arg[0] != '_')
 		return (false);
 	while (arg[i] != '=' && arg[i] != '\0')
 	{
-		if (ft_isalnum(arg[i]) == 0)
+		if (ft_isalnum(arg[i]) == 0 && arg[i] != '_')
 			return (false);
 		i++;
 	}
