--- src/skip_single_quote.c
@@ -1,10 +1,0 @@
-#include "../include/minishell.h"
-
-void	skip_single_quote(char *line)
-{
-	while (*line != '\'' && *line != '\0')
-	{
-		line++;
-	}
-	line++;
-}
