--- src/unwrap_dollar.c
@@ -1,25 +1,0 @@
-#include "../include/minishell.h"
-
-char	*unwrap_dollar(char *line)
-{
-	char	*new_line;
-	char	*value;
-
-	new_line = ft_calloc(10, 1);
-	while (*line)
-	{
-		if (*line == '\'')
-			skip_single_quote(line);
-		if (*line == '$')
-		{
-			value = return_dollar(line);
-			new_line = ft_strjoinfree(new_line, value);
-			free(value);
-			value = NULL;
-			line = skip_dollar(line);
-		}
-		new_line = charjoinfree(new_line, *line);
-		line++;
-	}
-	return (new_line);
-}
